TimeTable Generator release ver/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==19.1.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bitstring==4.0.1
blinker==1.5
bottle==0.12.25
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-bootstrap3==23.4
django-multiselectfield==0.1.12
ecdsa==0.18.0
entrypoints==0.4
esptool==4.5.1
et-xmlfile==1.1.0
filelock==3.10.7
firebase-admin==6.1.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.12.0
h8mail==2.5.6
holehe==1.61
httpcore==0.13.7
httplib2==0.21.0
httpx==1.0.0b0
idna==2.10
importlib-metadata==6.1.0
instagram-private-api==1.6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
maxminddb==2.2.0
mdurl==0.1.2
msgpack==1.0.4
multidict==6.0.4
mysql==0.0.3
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.2.0
prettytable==0.7.2
proto-plus==1.22.2
protobuf==3.20.3
proxy-tools==0.1.0
proxybroker==0.3.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyreadline==2.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywebview==4.0.2
reedsolo==1.6.0
requests==2.28.2
requests-futures==1.0.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.8.0
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stem==1.8.1
streamlit==1.20.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
torrequest==0.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
undetected-chromedriver==3.1.7
uritemplate==4.1.1
urllib3==1.26.14
users==1.0.dev0
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
webdriver-manager==3.8.2
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
